Add support for typing and referencing notes
This adds auto-generated citekeys for notes, so that any note
can be referenced in a manner similar to excerpts. It also
adds the additional yaml filed "type". A note citekey will be
"type:filename", where type defaults to "note" (if not set)
and filename is the filename of the markdown file stripped of
any leading directories and the ".md" extension.