Rewrite jotter-keyword in python
This replaces the initial shell script implementation of
jotter-keyword with an equivalent python implementation,
relying on the survey function from jotter.py. The rewrite
leads to significantly improved performance --- due to the
fact that extracting keywords from files had been implemented
in a rather inefficient way in the shell version.