Perform keyword operations relative to cwd
This changes the default behaviour of jotter-keyword.py
to only consider files in the current working directory
and its subdirectories for operations. I feel that this
is more useful while working than the previous behaviour
of searching the complete jotter. A command line option
has been added to get the old behaviour.