Add jotter-ln
This includes changes to the jotter file format. The '.jotter'
directory now includes a file named 'id' that contains a unique
identifier of a given jotter. This allows to prevent linking to the
same jotter multiple times. Furthermore, the '.jotter' directory
may now also include symlinks to other directories which will,
when they exist, be treated as belonging to the same jotter.